refactor(main): expose only functional api
fludic api is an overkill for this library
