--- fp.ts
@@ -1,28 +1,0 @@
-/**
- * Created by tushar.mathur on 11/10/16.
- */
-
-export {create} from './src/sources/Create'
-export {debounce} from './src/operators/Debounce'
-export {delay} from './src/operators/Delay'
-export {filter} from './src/operators/Filter'
-export {forEach} from './src/lib/ForEach'
-export {frames} from './src/sources/Frames'
-export {fromDOM} from './src/sources/FromDOM'
-export {fromPromise} from './src/sources/FromPromise'
-export {interval} from './src/sources/Interval'
-export {IObservable} from './src/lib/Observable'
-export {IObserver} from './src/lib/Observer'
-export {ISubscriberFunction} from './src/lib/SubscriberFunction'
-export {ISubscription} from './src/lib/Subscription'
-export {join, flatMap} from './src/operators/Join'
-export {map} from './src/operators/Map'
-export {merge} from './src/operators/Merge'
-export {multicast} from './src/operators/Multicast'
-export {of} from './src/sources/FromArray'
-export {reduce} from './src/operators/Reduce'
-export {sample} from './src/operators/Sample'
-export {scan} from './src/operators/Scan'
-export {skipRepeats} from './src/operators/SkipRepeats'
-export {slice} from './src/operators/Slice'
-export {switchLatest, switchMap} from './src/operators/Switch'
