+++ src/main.ts
@@ -2,9 +2,29 @@
  * Created by tushar on 17/02/17.
  */
 
+export {create} from './sources/Create'
+export {debounce} from './operators/Debounce'
+export {delay} from './operators/Delay'
+export {filter} from './operators/Filter'
+export {forEach} from './lib/ForEach'
+export {frames} from './sources/Frames'
+export {fromDOM} from './sources/FromDOM'
+export {fromPromise} from './sources/FromPromise'
+export {interval} from './sources/Interval'
 export {IObservable} from './lib/Observable'
 export {IObserver} from './lib/Observer'
 export {IScheduler} from './lib/Scheduler'
 export {ISubscriberFunction} from './lib/SubscriberFunction'
 export {ISubscription} from './lib/Subscription'
-export {Observable} from './lib/Observable'
+export {join, flatMap} from './operators/Join'
+export {map} from './operators/Map'
+export {merge} from './operators/Merge'
+export {multicast} from './operators/Multicast'
+export {of} from './sources/FromArray'
+export {reduce} from './operators/Reduce'
+export {sample} from './operators/Sample'
+export {scan} from './operators/Scan'
+export {skipRepeats} from './operators/SkipRepeats'
+export {slice} from './operators/Slice'
+export {switchLatest, switchMap} from './operators/Switch'
+export {fromArray} from './sources/FromArray'
